Moves SQL configration to the sql package
