+++ opsml/registry/sql/base/__init__.py
@@ -3,6 +3,7 @@
 # LICENSE file in the root directory of this source tree.
 
 from opsml.registry.storage.settings import settings
+from opsml.settings.config import config
 
 
 # no typing in order to prevent imports before they're needed
@@ -16,11 +17,13 @@ def get_registry():  # type: ignore
         from sqlalchemy.engine.base import Engine
 
         from opsml.registry.sql.base.server import ServerRegistry
+        from opsml.registry.sql.connectors.connector import DefaultConnector
         from opsml.registry.sql.db_initializer import DBInitializer
         from opsml.registry.sql.table_names import RegistryTableNames
 
+        connector = DefaultConnector(tracking_uri=config.opsml_tracking_uri).get_connector()
         db_initializer = DBInitializer(
-            engine=cast(Engine, settings.connection_client.sql_engine),
+            engine=cast(Engine, connector.sql_engine),
             registry_tables=list(RegistryTableNames),
         )
         db_initializer.initialize()
