--- opsml/registry/sql/connectors/__init__.py
@@ -1,2 +1,0 @@
-from opsml.registry.sql.connectors.base import BaseSQLConnection  # noqa: F401
-from opsml.registry.sql.connectors.connector import SQLConnector  # noqa: F401
